Rename check_sdk_imports to check_sdk_imports_in_core
Mirrors the naming of the sibling check_core_imports_in_sdk hook so the
direction of each guard (what's checked in where) is obvious from the
filename.
